test(react): Ensure react router 3 tests always have correct types
