+++ packages/react/package.json
@@ -54,9 +54,8 @@
     "@types/hoist-non-react-statics": "^3.3.5",
     "@types/node-fetch": "^2.6.11",
     "@types/react": "17.0.3",
-    "@types/react-router-3": "npm:@types/react-router@3.0.24",
-    "@types/react-router-4": "npm:@types/react-router@5.1.14",
-    "@types/react-router-5": "npm:@types/react-router@5.1.14",
+    "@types/react-router-4": "npm:@types/react-router@4.0.25",
+    "@types/react-router-5": "npm:@types/react-router@5.1.20",
     "eslint-plugin-react": "^7.20.5",
     "eslint-plugin-react-hooks": "^4.0.8",
     "history-4": "npm:history@4.6.0",
